Adding a missing input file
I am not sure what to call it, but this is a required input file, just under the hood, because it should be compatible with the masses defined in the CHARMM 36 parameterizations. Please feel free to update and verify this information. It is a list of dictionaries that define a variety of strings, allowing it to be used in a few ways. I lost the file that generated the pickle, though. Just reach out to me and I can help reconstruct it if it is a problem/want to recreate it.